Migrate stars component to styled-component
