--- src/stars/style.ts
@@ -1,8 +1,0 @@
-import css from 'styled-jsx/css'
-
-export default css`
-  .star {
-    display: inline-block;
-    margin-right: 2px;
-  }
-`
